Refactor accordion filter styles
