+++ context/app/static/js/components/Search/style.js
@@ -1,1 +1,9 @@
-export * from './styleForSidebar';
+import styled from 'styled-components';
+
+import { SideBar } from 'searchkit';
+
+const StyledSideBar = styled(SideBar)`
+  padding: 0;
+`;
+
+export { StyledSideBar };
